requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
jinja2
requests
